pyproject.toml
[project]
name = "pystum3-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Gooey>=1.0.8",
    "pystun3>=2.0.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.12.0",
]
